- db first ile tablolar hem codda hem databasede oluşturuldu. - controllarda get ve post methodu yazıldı - get methodu postman ve angular tarafından test edildi. - angular dan gelen isteklerin  cross politikalarına takılmaması için back-end de gerekli düzenlemeler gerçekleştirildi.

+++ Core/ETicaretAPI.Domain/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace ETicaretAPI.Domain.Entities
     {
         public string Name { get; set; }
         public int Stock { get; set; }
-        public int Price { get; set; }
+        public float Price { get; set; }
         public ICollection<Order> Orders { get; set; } //Productın birden çok orderı olabiliyor. == mantığı order sayfasında entites de açıkladım.
     }
 }
